# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MoobilPay API",
        "description": "API complète de paiement en ligne avec Stripe et PayPal. Permet de gérer les paiements, méthodes de paiement et webhooks de manière sécurisée.",
        "contact": {
            "name": "Support MoobilPay",
            "email": "tchindavaldoblair@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.moobilpay.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Profil utilisateur",
                "description": "Récupère les informations de l'utilisateur authentifié",
                "operationId": "getCurrentUser",
                "responses": {
                    "200": {
                        "description": "Informations utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "john@example.com"
                                        },
                                        "email_verified_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-11-14T10:30:00Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-11-14T10:30:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-11-14T10:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/webhooks/stripe": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Webhook Stripe",
                "description": "Réception des événements de paiement Stripe",
                "operationId": "45cf3a5e72c504da57fec6e857eec011",
                "requestBody": {
                    "description": "Payload du webhook Stripe",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "evt_1234567890"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "payment_intent.succeeded"
                                    },
                                    "data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook traité avec succès"
                    },
                    "400": {
                        "description": "Erreur de signature ou payload invalide"
                    }
                }
            }
        },
        "/api/webhooks/paypal": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Webhook PayPal",
                "description": "Réception des événements de paiement PayPal",
                "operationId": "618ebbefd07820ee18d76b6c46c9ca24",
                "requestBody": {
                    "description": "Payload du webhook PayPal",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "WH-1234567890"
                                    },
                                    "event_type": {
                                        "type": "string",
                                        "example": "PAYMENT.CAPTURE.COMPLETED"
                                    },
                                    "resource": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook traité avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation du webhook"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum. Utilisez votre token d'API.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}